src/requirements.txt
datasets==2.10.1
evaluate==0.4.0
highlight-text==0.2
matplotlib==3.7.1
openai==2.7.1
scikit-learn==1.2.2
tiktoken==0.3.3
torch==2.0.0
transformers==4.26.1
wandb==0.14.0
python-dotenv